So many things...
I just went along and refactored a bunch of code. The first step was standardising the elements (Card, Grid, Headings, etc.) so that they are centrally defined and locally modified when needed via props. This resulted in deleting many files with mostly repeated code and introducing a few ones. While doing this, most of the components were switched to their functional style (instead of classes) to make them better suited to composability and clarity. Animations are now fetched from the API instead of being locally managed. There is also a new 'session' object, with it's state managed by React to pass along the previous answer to the next. This is a first step in trying to be able to have several chained questions independent of this app.